Update human-summary printers
