+++ tealer/__main__.py
@@ -63,21 +63,18 @@ To choose printers to run from the available list:
 """
 
 import argparse
-import inspect
 import re
 import sys
 import json
 from pathlib import Path
 from typing import List, Any, Type, Tuple, TYPE_CHECKING, Optional
 
-from pkg_resources import iter_entry_points, require  # type: ignore
+from pkg_resources import require  # type: ignore
 
-from tealer.detectors import all_detectors
 from tealer.detectors.abstract_detector import AbstractDetector, DetectorType
-from tealer.printers import all_printers
 from tealer.printers.abstract_printer import AbstractPrinter
 from tealer.teal.parse_teal import parse_teal
-from tealer.utils.command_line import output_detectors, output_printers
+from tealer.utils.command_line import get_detectors_and_printers, output_detectors, output_printers
 from tealer.utils.output import cfg_to_dot
 from tealer.utils.algoexplorer import (
     get_application_using_app_id,
@@ -351,74 +348,6 @@ class ListPrinters(argparse.Action):  # pylint: disable=too-few-public-methods
         _, printers = get_detectors_and_printers()
         output_printers(printers)
         parser.exit()
-
-
-def collect_plugins() -> Tuple[List[Type[AbstractDetector]], List[Type[AbstractPrinter]]]:
-    """collect detectors and printers installed in form of plugins.
-
-    plugins are collected using the entry point group `teal_analyzer.plugin`.
-    The entry point of each plugin has to return tuple containing list of detectors and
-    list of printers defined in the plugin when called.
-
-    Returns:
-        (Tuple[List[Type[AbstractDetector]], List[Type[AbstractPrinter]]]): detectors and
-        printers added in the form of plugins.
-
-    """
-    detector_classes: List[Type[AbstractDetector]] = []
-    printer_classes: List[Type[AbstractPrinter]] = []
-    for entry_point in iter_entry_points(group="teal_analyzer.plugin", name=None):
-        make_plugin = entry_point.load()
-
-        plugin_detectors, plugin_printers = make_plugin()
-        detector = None
-        if not all(issubclass(detector, AbstractDetector) for detector in plugin_detectors):
-            raise TealerException(
-                f"Error when loading plugin {entry_point}, {detector} is not a detector"
-            )
-        printer = None
-        if not all(issubclass(printer, AbstractPrinter) for printer in plugin_printers):
-            raise TealerException(
-                f"Error when loading plugin {entry_point}, {printer} is not a printer"
-            )
-
-        detector_classes += plugin_detectors
-        printer_classes += plugin_printers
-
-    return detector_classes, printer_classes
-
-
-# from slither: slither/__main__.py
-def get_detectors_and_printers() -> Tuple[
-    List[Type[AbstractDetector]], List[Type[AbstractPrinter]]
-]:
-    """Get list of detectors and printers available to tealer.
-
-    Detectors, Printers are considered available to tealer either if they
-    are defined in the tealer itself or if they are defined in one of the
-    tealer plugins installed in the system.
-
-    Returns:
-        list of detectors, list of printers defined in tealer and plugins
-        combined.
-    """
-
-    detector_classes = [getattr(all_detectors, name) for name in dir(all_detectors)]
-    detector_classes = [
-        d for d in detector_classes if inspect.isclass(d) and issubclass(d, AbstractDetector)
-    ]
-
-    printer_classes = [getattr(all_printers, name) for name in dir(all_printers)]
-    printer_classes = [
-        d for d in printer_classes if inspect.isclass(d) and issubclass(d, AbstractPrinter)
-    ]
-
-    plugins_detectors, plugins_printers = collect_plugins()
-
-    detector_classes += plugins_detectors
-    printer_classes += plugins_printers
-
-    return detector_classes, printer_classes
 
 
 def handle_print_cfg(args: argparse.Namespace, teal: "Teal") -> None:
