+++ tealer/printers/human_summary.py
@@ -17,6 +17,7 @@ from typing import List, Tuple, Optional, TYPE_CHECKING
 from tealer.printers.abstract_printer import AbstractPrinter
 from tealer.teal.instructions.instructions import Label, contract_type_to_txt
 from tealer.utils.code_complexity import compute_cyclomatic_complexity
+from tealer.detectors.abstract_detector import DetectorClassification
 
 if TYPE_CHECKING:
     from pathlib import Path
@@ -43,6 +44,7 @@ class PrinterHumanSummary(AbstractPrinter):
         result = "Yes" if is_complex else "No"
         return result
 
+    # pylint: disable=too-many-locals
     def _get_detectors_result(self) -> Tuple[List, int, int, int, int, int]:
         """Return dectector results and number of issues found in the contract for each `impact` class.
 
@@ -52,25 +54,36 @@ class PrinterHumanSummary(AbstractPrinter):
             the same order.
 
         """
-        checks_optimization = self.teal.detectors_optimization
-        checks_informational = self.teal.detectors_informational
-        checks_low = self.teal.detectors_low
-        checks_medium = self.teal.detectors_medium
-        checks_high = self.teal.detectors_high
+        from tealer.utils.command_line import (  # pylint: disable=import-outside-toplevel
+            get_detectors_and_printers,
+        )
 
-        issues_optimization = [c.detect() for c in checks_optimization]
+        detector_classes, _ = get_detectors_and_printers()
+        detectors = [d(self.teal) for d in detector_classes]
+
+        checks_optimization = [
+            d for d in detectors if d.IMPACT == DetectorClassification.OPTIMIZATION
+        ]
+        checks_informational = [
+            d for d in detectors if d.IMPACT == DetectorClassification.INFORMATIONAL
+        ]
+        checks_low = [d for d in detectors if d.IMPACT == DetectorClassification.LOW]
+        checks_medium = [d for d in detectors if d.IMPACT == DetectorClassification.MEDIUM]
+        checks_high = [d for d in detectors if d.IMPACT == DetectorClassification.HIGH]
+
+        issues_optimization = [c.detect().paths for c in checks_optimization]
         issues_optimization = [c for c in issues_optimization if c]
 
-        issues_informational = [c.detect() for c in checks_informational]
+        issues_informational = [c.detect().paths for c in checks_informational]
         issues_informational = [c for c in issues_informational if c]
 
-        issues_low = [c.detect() for c in checks_low]
+        issues_low = [c.detect().paths for c in checks_low]
         issues_low = [c for c in issues_low if c]
 
-        issues_medium = [c.detect() for c in checks_medium]
+        issues_medium = [c.detect().paths for c in checks_medium]
         issues_medium = [c for c in issues_medium if c]
 
-        issues_high = [c.detect() for c in checks_high]
+        issues_high = [c.detect().paths for c in checks_high]
         issues_high = [c for c in issues_high if c]
 
         all_results = (
